requirements.txt
imapclient
google-generativeai
pyyaml
beautifulsoup4